+++ src/core/control/CrashHandlerUnix.h
@@ -141,7 +141,7 @@ static void crashHandler(int sig) {
        << std::endl
        << std::endl;
 
-    messages = backtrace_symbols(array, size);
+    messages = backtrace_symbols(array, static_cast<int>(size));
 
     for (size_t i = 0; i < size; i++) {
         fp << FORMAT_STR("[bt]: ({1}) {2}") % i % messages[i];
