+++ src/core/gui/inputdevices/HandRecognition.cpp
@@ -115,7 +115,7 @@ auto HandRecognition::enableTimeout(HandRecognition* self) -> bool {
         return false;
     }
 
-    auto nextTime = now - self->lastPenAction + self->disableTimeout;
+    auto nextTime = static_cast<guint>(now - self->lastPenAction + self->disableTimeout);
 
     g_timeout_add(nextTime, xoj::util::wrap_v<enableTimeout>, self);
